move base64 ref code into kapitan/refs/base64.py, fix tests
