+++ kapitan/refs/secrets/awskms.py
@@ -17,7 +17,8 @@
 import base64
 import boto3
 
-from kapitan.refs.base import Base64Ref, Base64RefBackend, RefError
+from kapitan.refs.base import RefError
+from kapitan.refs.base64 import Base64Ref, Base64RefBackend
 from kapitan import cached
 from kapitan.errors import KapitanError
