+++ kapitan/refs/secrets/gpg.py
@@ -19,7 +19,8 @@ import gnupg
 import logging
 import time
 
-from kapitan.refs.base import Base64Ref, Base64RefBackend, RefError
+from kapitan.refs.base import RefError
+from kapitan.refs.base64 import Base64Ref, Base64RefBackend
 from kapitan import cached
 from kapitan.errors import KapitanError
